requirements.txt
absl-py==2.3.1
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
autopage==0.5.2
av==16.0.1
babel==2.17.0
backports.strenum==1.3.1
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cliff==4.13.1
cmaes==0.12.0
cmd2==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
conformer==0.3.2
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
csvw==3.7.0
cycler==0.12.1
Cython==3.2.4
datasets==2.18.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.36.0
dill==0.3.8
distlib==0.4.0
distro==1.9.0
dlinfo==2.0.0
duckduckgo_search==8.1.1
edge-tts==6.1.17
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpeg==1.4
ffmpy==1.0.0
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.2.0
funasr==1.1.12
gast==0.7.0
gdown==5.2.0
gradio==3.43.2
gradio_client==0.5.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
HyperPyYAML==1.2.2
identify==2.6.15
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.4.5
inflect==7.5.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.38.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jaconv==0.4.1
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaldiio==2.18.1
kiwisolver==1.4.9
langchain==1.2.3
langchain-core==1.2.7
langdetect==1.0.9
langfuse==4.5.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langid==1.1.6
langsmith==0.6.2
language-tags==1.2.0
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matcha-tts==0.0.7.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
modelscope==1.15.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.10.0
notebook==7.5.1
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.19.2
openai==2.15.0
openai-whisper==20231117
opencv-python==4.10.0.84
optuna==2.10.1
orjson==3.11.5
ormsgpack==1.12.1
oss2==2.19.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
phonemizer==3.3.0
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.5.1
prettytable==3.17.0
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
PyAudio==0.2.14
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pymilvus==2.6.6
pynini @ file:///D:/bld/pynini_1756319225364/work
pynndescent==0.6.0
pyparsing==3.3.1
pyperclip==1.11.0
pypinyin==0.55.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorch-wpe==0.0.1
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.8
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.2
rootutils==1.0.7
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
segments==2.3.0
semantic-version==2.10.0
Send2Trash==2.0.0
sentencepiece==0.2.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
stevedore==5.6.0
sympy==1.14.0
tavily==1.1.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.3.0
torch @ file:///D:/Pytorch%E5%AE%89%E8%A3%85%E6%96%87%E4%BB%B6/torch-2.8.0%2Bcu128-cp310-cp310-win_amd64.whl#sha256=43938e9a174c90e5eb9e906532b2f1e21532bbfa5a61b65193b4f54714d34f9e
torch-complex==0.4.4
torchaudio @ file:///D:/Pytorch%E5%AE%89%E8%A3%85%E6%96%87%E4%BB%B6/torchaudio-2.8.0%2Bcu128-cp310-cp310-win_amd64.whl#sha256=5d7a9d913e2744573ed3b7ec2f781ed39833c81c9c41859973ec10ac174c2366
torchmetrics==1.8.2
torchvision @ file:///D:/Pytorch%E5%AE%89%E8%A3%85%E6%96%87%E4%BB%B6/torchvision-0.23.0%2Bcu128-cp310-cp310-win_amd64.whl#sha256=8ec6f2281ef5d52471b01b99eb04243d0c2cccb1972ba43217085025fe5a6c3f
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.9.post2
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==20.36.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==11.0.3
Werkzeug==3.1.5
WeTextProcessing==1.0.4.1
wget==3.2
widgetsnbextension==4.0.15
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
